mv typeorm config
