+++ package.json
@@ -23,9 +23,9 @@
     "test:e2e": "jest --config ./test/jest-e2e.json",
     "seed": "bash ./staging_backup.sh -c",
     "typeorm": "typeorm-ts-node-commonjs",
-    "migration:generate": "yarn typeorm -- -d ./typeorm.config.ts migration:generate ./migrations/bloom_backend",
-    "migration:run": "yarn typeorm migration:run -- -d ./typeorm.config.ts",
-    "migration:revert": "yarn typeorm -- -d ./typeorm.config.ts migration:revert",
+    "migration:generate": "yarn typeorm -- -d ./src/typeorm.config.ts migration:generate ./src/migrations/bloom-backend",
+    "migration:run": "yarn typeorm migration:run -- -d ./src/typeorm.config.ts",
+    "migration:revert": "yarn typeorm -- -d ./src/typeorm.config.ts migration:revert",
     "migration:show": "yarn build && yarn typeorm -- migration:show"
   },
   "dependencies": {
